@@ -6,14 +6,14 @@
 
     <groupId>com.stanexe</groupId>
     <artifactId>PlayerReferrals</artifactId>
-    <version>1.3.3</version>
+    <version>1.4.0</version>
     <packaging>jar</packaging>
 
     <name>PlayerReferrals</name>
 
     <description>Allow your players to get rewards for inviting their friends!</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>21</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
     <url>https://stanexe.com</url>
