@@ -17,11 +17,13 @@ public class ReferralLeaderboardCommand implements CommandExecutor {
 
     final PlayerReferrals plugin = PlayerReferrals.getInstance();
     final FileConfiguration messages = plugin.getMessagesConfig();
+
     @Override
     public boolean onCommand(@NotNull CommandSender commandSender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
         if (!plugin.getConfig().getBoolean("enable-leaderboard")) {
             commandSender.sendMessage(ChatColor.RED + "The leaderboard is disabled.");
         }
+
         List<RefUser> topPlayers = DatabaseUtil.getTopPlayers();
         StringBuilder message = new StringBuilder(ChatColor.DARK_AQUA + "Referral Leaderboard:");
         int i = 1;
@@ -32,9 +34,9 @@ public class ReferralLeaderboardCommand implements CommandExecutor {
                 message.append("\n").append(i).append(". ").append(Bukkit.getOfflinePlayer(user.getUUID()).getName()).append(" - ").append(user.getPlayerScore());
                 i++;
             }
+
             commandSender.sendMessage(message.toString());
         }
-
 
         return true;
     }
