@@ -8,9 +8,8 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.UUID;
 
-
-
 public class PlayerReferralsExpansion extends PlaceholderExpansion {
+
     private final PlayerReferrals plugin = PlayerReferrals.getInstance();
     @Override
     public boolean canRegister() {
@@ -35,36 +34,34 @@ public class PlayerReferralsExpansion extends PlaceholderExpansion {
     @Override
     public String onRequest(OfflinePlayer player, String identifier) {
 
-        if (identifier.equals("score")) {
-            RefUser refUser = new RefUser(player.getUniqueId());
-            return String.valueOf(refUser.getPlayerScore());
-        }
-
-        if (identifier.equals("referrerign")) {
-            RefUser refUser = new RefUser(player.getUniqueId());
-            UUID uuid = refUser.getReferrer();
-            if (uuid != null) {
-                return Bukkit.getOfflinePlayer(uuid).getName();
-            } else {
-                String text = plugin.getMessagesConfig().getString("referrerign-placeholder-empty");
-                if (text == null) {
-                    return "";
+        switch (identifier) {
+            case "score": {
+                RefUser refUser = new RefUser(player.getUniqueId());
+                return String.valueOf(refUser.getPlayerScore());
+            }
+            case "referring": {
+                RefUser refUser = new RefUser(player.getUniqueId());
+                UUID uuid = refUser.getReferrer();
+                if (uuid != null) {
+                    return Bukkit.getOfflinePlayer(uuid).getName();
+                } else {
+                    String text = plugin.getMessagesConfig().getString("referring-placeholder-empty");
+                    if (text == null) {
+                        return "";
+                    }
+                    return StringTools.colors(text);
                 }
-                return StringTools.colors(text);
             }
-
-        }
-
-        if (identifier.equals("minutesremaining")) {
-            RefUser refUser = new RefUser(player.getUniqueId());
-            long minutes = refUser.getMinutesRemaining();
-            if (minutes < 0) {
-                minutes = 0;
+            case "minutesremaining": {
+                RefUser refUser = new RefUser(player.getUniqueId());
+                long minutes = refUser.getMinutesRemaining();
+                if (minutes < 0) {
+                    minutes = 0;
+                }
+                return String.valueOf(minutes);
             }
-            return String.valueOf(minutes);
         }
 
         return null;
     }
-
 }
